+++ CHANGELOG.md
@@ -1,9 +1,8 @@
 # Road map
 
-- [ ] Comment/uncomment
-- [ ] Brace matching
+- [x] Comment/uncomment
+- [x] Brace matching
 - [ ] Implement custom parser
-- [ ] Formatting command
 - [ ] Formatting command
 
 Features that have a checkmark are complete and available for